+++ packages/api-services/src/services/legal_page/index.ts
@@ -29,6 +29,7 @@ export class LegalPageApiClient {
     const p = urlParams(params);
     return apiRequest({...config, url: `${Endpoints.GetLegalPages.url({})}?${p.toString()}`, method: Endpoints.GetLegalPages.method});
   }
+
   /**
    * This endpoint retrieves a specific legal page.
    * @param id
@@ -37,6 +38,7 @@ export class LegalPageApiClient {
   static getSpecificLegalPage(id: number | string, config?: AxiosRequestConfig): Promise<SCLegalPageType> {
     return apiRequest({...config, url: Endpoints.LegalPage.url({id}), method: Endpoints.LegalPage.method});
   }
+
   /**
    * This endpoint retrieves all revisions of a legal page.
    * @param id
@@ -44,6 +46,13 @@ export class LegalPageApiClient {
    */
   static getAllRevisionsOfLegalPage(policy: string, config?: AxiosRequestConfig): Promise<SCPaginatedResponse<SCLegalPageType>> {
     return apiRequest({...config, url: Endpoints.LegalPageRevisions.url({policy}), method: Endpoints.LegalPageRevisions.method});
+  }
+
+  /**
+   * This endpoint retrieves all last revisions of legal pages.
+   */
+  static getAllLastRevisionsOfLegalPages(config?: AxiosRequestConfig): Promise<SCLegalPageType[]> {
+    return apiRequest({...config, url: Endpoints.LegalPagesLastRevision.url({}), method: Endpoints.LegalPagesLastRevision.method});
   }
 
   /**
@@ -71,7 +80,7 @@ export class LegalPageApiClient {
    * @param accept Accept or not accept a legal page, valid values are: ('true', 'on', '1').
    * @param config
    */
-  static ackLegalPage(id: number | string, accept?: number, config?: AxiosRequestConfig): Promise<SCLegalPageAckType> {
+  static ackLegalPage(id: number | string, accept?: number | string, config?: AxiosRequestConfig): Promise<SCLegalPageAckType> {
     return apiRequest({...config, url: Endpoints.AckLegalPage.url({id}), method: Endpoints.AckLegalPage.method, data: {accept: accept} ?? null});
   }
 
@@ -140,6 +149,10 @@ export default class LegalPageService {
     return LegalPageApiClient.getAllRevisionsOfLegalPage(policy, config);
   }
 
+  static async getAllLastRevisionsOfLegalPages(config?: AxiosRequestConfig): Promise<SCLegalPageType[]> {
+    return LegalPageApiClient.getAllLastRevisionsOfLegalPages(config);
+  }
+
   static async getLastRevisionOfLegalPage(policy: string, config?: AxiosRequestConfig): Promise<SCLegalPageType> {
     return LegalPageApiClient.getLastRevisionOfLegalPage(policy, config);
   }
@@ -148,7 +161,7 @@ export default class LegalPageService {
     return LegalPageApiClient.searchLegalPages(params, config);
   }
 
-  static async ackLegalPage(id: number | string, accept?: number, config?: AxiosRequestConfig): Promise<SCLegalPageAckType> {
+  static async ackLegalPage(id: number | string, accept?: number | string, config?: AxiosRequestConfig): Promise<SCLegalPageAckType> {
     return LegalPageApiClient.ackLegalPage(id, accept, config);
   }