+++ packages/react-ui/src/components/ConsentSolution/index.tsx
@@ -1,0 +1,4 @@
+import ConsentSolution, {ConsentSolutionProps} from './ConsentSolution';
+import ConsentSolutionSkeleton from './Skeleton';
+export default ConsentSolution;
+export {ConsentSolutionProps, ConsentSolutionSkeleton};