+++ packages/react-ui/src/components/Footer/Footer.tsx
@@ -1,10 +1,10 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect, useMemo, useState} from 'react';
 import {useThemeProps} from '@mui/system';
 import {styled} from '@mui/material/styles';
 import {Box, Grid, Typography} from '@mui/material';
 import classNames from 'classnames';
 import {http, Endpoints, HttpResponse} from '@selfcommunity/api-services';
-import {SCLegalPagePoliciesType} from '@selfcommunity/types';
+import {SCCustomAdvPosition, SCLegalPagePoliciesType} from '@selfcommunity/types';
 import {Logger} from '@selfcommunity/utils';
 import {
   SCRoutes,
@@ -13,11 +13,14 @@ import {
   useSCRouting,
   SCPreferences,
   SCPreferencesContextType,
-  useSCPreferences
+  useSCPreferences,
+  SCUserContextType,
+  useSCUser
 } from '@selfcommunity/react-core';
 import {SCOPE_SC_UI} from '../../constants/Errors';
 import FooterSkeleton from './Skeleton';
 import {FormattedMessage} from 'react-intl';
+import CustomAdv from '../CustomAdv';
 
 const PREFIX = 'SCFooter';
 
@@ -57,6 +60,8 @@ export interface FooterProps {
   [p: string]: any;
 }
 
+const PREFERENCES = [SCPreferences.ADVERTISING_CUSTOM_ADV_ENABLED, SCPreferences.ADVERTISING_CUSTOM_ADV_ONLY_FOR_ANONYMOUS_USERS_ENABLED];
+
 /**
  *> API documentation for the Community-JS Footer component. Learn about the available props and the CSS API.
  *
@@ -79,7 +84,7 @@ export interface FooterProps {
  * @param inProps
  */
 export default function Footer(inProps: FooterProps): JSX.Element {
-  //PROPS
+  // PROPS
   const props: FooterProps = useThemeProps({
     props: inProps,
     name: PREFIX
@@ -88,14 +93,42 @@ export default function Footer(inProps: FooterProps): JSX.Element {
 
   // CONTEXT
   const scRoutingContext: SCRoutingContextType = useSCRouting();
+  const scUserContext: SCUserContextType = useSCUser();
 
   // PREFERENCES
   const scPreferences: SCPreferencesContextType = useSCPreferences();
   const copyRight = scPreferences.preferences[SCPreferences.TEXT_APPLICATION_COPYRIGHT].value;
 
-  //STATE
+  // STATE
   const [pages, setPages] = useState<any[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
+
+  /**
+   * Compute preferences
+   */
+  const preferences = useMemo(() => {
+    const _preferences = {};
+    PREFERENCES.map((p) => (_preferences[p] = p in scPreferences.preferences ? scPreferences.preferences[p].value : null));
+    return _preferences;
+  }, [scPreferences.preferences]);
+
+  /**
+   * Render advertising
+   */
+  function renderAdvertising() {
+    if (
+      preferences[SCPreferences.ADVERTISING_CUSTOM_ADV_ENABLED] &&
+      ((preferences[SCPreferences.ADVERTISING_CUSTOM_ADV_ONLY_FOR_ANONYMOUS_USERS_ENABLED] && scUserContext.user === null) ||
+        !preferences[SCPreferences.ADVERTISING_CUSTOM_ADV_ONLY_FOR_ANONYMOUS_USERS_ENABLED])
+    ) {
+      return (
+        <Grid item>
+          <CustomAdv position={SCCustomAdvPosition.POSITION_ABOVE_FOOTER_BAR} />
+        </Grid>
+      );
+    }
+    return null;
+  }
 
   /**
    * Fetches custom pages
@@ -136,6 +169,7 @@ export default function Footer(inProps: FooterProps): JSX.Element {
   return (
     <Root {...rest} className={classNames(classes.root, className)}>
       <Grid container spacing={1} justifyContent="center">
+        {renderAdvertising()}
         {pages.map((page, index) => (
           <Grid item key={index}>
             <Link