+++ packages/react-ui/src/components/UserProfileInfo/UserProfileInfo.stories.tsx
@@ -11,11 +11,11 @@ export default {
     userId: {
       control: {type: 'number'},
       description: 'User Id',
-      table: {defaultValue: {summary: 1}}
+      table: {defaultValue: {summary: 7}}
     }
   },
   args: {
-    userId: 1
+    userId: 7
   }
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
 } as ComponentMeta<typeof UserProfileInfo>;