+++ packages/react-ui/src/components/UserProfileInfo/UserProfileInfo.tsx
@@ -10,6 +10,8 @@ import UserProfileInfoSkeleton from './Skeleton';
 import classNames from 'classnames';
 import {useThemeProps} from '@mui/system';
 import {SCUserProfileFields} from '../../types';
+import TagChip from '../../shared/TagChip';
+import {isArray} from 'lodash';
 
 const messages = defineMessages({
   realName: {
@@ -41,6 +43,10 @@ const messages = defineMessages({
     defaultMessage: 'ui.userProfileInfo.gender'
   },
   website: {
+    id: 'ui.userProfileInfo.website',
+    defaultMessage: 'ui.userProfileInfo.website'
+  },
+  tags: {
     id: 'ui.userProfileInfo.website',
     defaultMessage: 'ui.userProfileInfo.website'
   }
@@ -138,6 +144,16 @@ export default function UserProfileInfo(inProps: UserProfileInfoProps): JSX.Elem
       case SCUserProfileFields.DATE_JOINED:
       case SCUserProfileFields.DATE_OF_BIRTH:
         return `${intl.formatDate(user[field], {year: 'numeric', month: 'numeric', day: 'numeric'})}`;
+      case SCUserProfileFields.TAGS:
+        return (
+          <>
+            {user.tags
+              .filter((t) => t.visible)
+              .map((t, i) => (
+                <TagChip tag={t} key={i} clickable={false} disposable={false} />
+              ))}
+          </>
+        );
       default:
         return user[field];
     }
@@ -156,6 +172,17 @@ export default function UserProfileInfo(inProps: UserProfileInfoProps): JSX.Elem
       <Grid container spacing={2}>
         {fields.map((field) => {
           if (scUser[field]) {
+            if (field === SCUserProfileFields.TAGS) {
+              return (
+                <React.Fragment key={field}>
+                  {scUser[field].length > 0 && (
+                    <Grid item xs={6}>
+                      {renderField(scUser, field)}
+                    </Grid>
+                  )}
+                </React.Fragment>
+              );
+            }
             return (
               <Grid item xs={6} key={field}>
                 <Typography variant="body2">