+++ packages/react-ui/src/constants/UserProfile.ts
@@ -1,6 +1,7 @@
 import {SCUserProfileFields, SCUserProfileSettings} from '../types';
 
 export const DEFAULT_FIELDS = [
+  SCUserProfileFields.TAGS,
   SCUserProfileFields.REAL_NAME,
   SCUserProfileFields.DATE_JOINED,
   SCUserProfileFields.DATE_OF_BIRTH,